refactor(math): move common.h integer-math macros into include/math.h
min/max, min_t/max_t, DIV_ROUND_UP, DIV_ROUND_CLOSEST, __round_mask and
round_up/round_down were general-purpose integer-math macros living in
include/common.h. Move them next to the other helpers in include/math.h and
have common.h include <math.h> (within its non-assembly section) so every
existing user keeps seeing them unchanged.

Non-math helpers (container_of, typecheck, BUG/ASSERT, panic, ...) stay in
common.h.
